refactor(MyInfoNotApplicable): push down MyInfoNotApplicable type only to the fields that require it
If a MyInfo field is not applicable (determined by source being set to the character value '3', then
it should be ignored completely without further processing.

Whether a field requires the NotApplicable type was determined by the data catalog hosted at the link below:

https://www.ndi-api.gov.sg/library/myinfo/implementation-myinfo-data

A field is potentially NotApplicable if its row in the catalog does not include all three SG/PR/FIN, i.e.
type determination for NotApplicable does not depend on whether the data is GovtVerified or UserProvided.

Applicable to SG citizens or PRs only:
       -  address
       -  aliasname
       -  countryofmarriage
       -  cpfbalances
       -  cpfcontributions
       -  cpfemployers
       -  divorcedate
       -  hanyupinyinaliasname
       -  hanyupinyinname
       -  hdbownership
       -  hdbtype
       -  housingtype
       -  marriagecertno
       -  marriedname
       -  residentialstatus
       -  secondaryrace
       -  sponsoredchildrenrecords

Applicable to FIN (foreigners) only:
       -  employmentsector
       -  housingtype
       -  passexpirydate
       -  passstatus
       -  passtype

Refactored to lift array types
       -  childrenbirthrecords